+++ packages/language-service/src/BUILD.bazel
@@ -3,7 +3,7 @@ load("//tools:defaults.bzl", "ts_library")
 package(default_visibility = ["//packages/language-service:__subpackages__"])
 
 ts_library(
-    name = "ivy",
+    name = "src",
     srcs = glob(["*.ts"]),
     deps = [
         "//packages/compiler",
