+++ packages/language-service/src/compiler_factory.ts
