+++ packages/language-service/src/completions.ts
