+++ packages/language-service/test/BUILD.bazel
@@ -12,8 +12,8 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/file_system/testing",
         "//packages/compiler-cli/src/ngtsc/testing",
         "//packages/compiler-cli/src/ngtsc/typecheck/api",
-        "//packages/language-service/ivy",
-        "//packages/language-service/ivy/testing",
+        "//packages/language-service/src",
+        "//packages/language-service/testing",
         "@npm//typescript",
     ],
 )
