+++ packages/language-service/test/diagnostic_spec.ts
