+++ packages/language-service/test/get_template_location_for_component_spec.ts
