+++ packages/language-service/test/legacy/project/app/#inner/component.ts
