+++ packages/language-service/test/legacy/project/app/#inner/inner.html
