+++ packages/language-service/test/legacy/project/app/app.component.ts
