+++ packages/language-service/test/legacy/project/app/main.ts
