+++ packages/language-service/test/legacy/template_target_spec.ts
@@ -10,8 +10,8 @@ import {ParseError, parseTemplate} from '@angular/compiler';
 import * as e from '@angular/compiler/src/expression_parser/ast';  // e for expression AST
 import * as t from '@angular/compiler/src/render3/r3_ast';         // t for template AST
 
-import {getTargetAtPosition, SingleNodeTarget, TargetNodeKind, TwoWayBindingContext} from '../../template_target';
-import {isExpressionNode, isTemplateNode} from '../../utils';
+import {getTargetAtPosition, SingleNodeTarget, TargetNodeKind, TwoWayBindingContext} from '../../src/template_target';
+import {isExpressionNode, isTemplateNode} from '../../src/utils';
 
 interface ParseResult {
   nodes: t.Node[];
